+++ src/index.jsx
@@ -21,9 +21,6 @@ import { detectColorsFromHTML } from 'helpers/colors'
 // 强化图片尺寸编辑功能
 // 增加取色器
 // 尝试支持mention功能
-// 允许插入相册功能
-// 增加插入水平分割线功能
-// 增加媒体文件的删除回调
 
 const editorDecorators = new CompositeDecorator(decorators)
 const blockRenderMap = DefaultDraftBlockRenderMap.merge(customBlockRenderMap)
@@ -88,6 +85,8 @@ export default class BraftEditor extends EditorController {
     if (!this.contentInitialized) {
       this.contentInitialized = true
       this.setContent(nextProps.initialContent)
+    } else if (nextProps.contentId !== this.props.contentId) {
+      this.setContent(nextProps.initialContent)
     }
 
   }
@@ -124,9 +123,9 @@ export default class BraftEditor extends EditorController {
   getContent = (format) => {
 
     format = format || this.props.contentFormat || 'raw'
+
     const contentState = this.contentState
-    let { fontFamilies} = this.props
-    fontFamilies = fontFamilies || defaultOptions.fontFamilies
+    const { fontFamilies} = this.props
 
     return format === 'html' ? convertToHTML(getToHTMLConfig({
       contentState, fontFamilies
@@ -155,14 +154,12 @@ export default class BraftEditor extends EditorController {
     let convertedContent
     let newState = {}
     let { contentFormat, colors, fontFamilies} = this.props
-    fontFamilies = fontFamilies || defaultOptions.fontFamilies
-    const presetColors = colors || defaultOptions.colors
 
     contentFormat = format || contentFormat || 'raw'
     if (contentFormat === 'html') {
-      content = content
-      newState.tempColors = [...this.state.tempColors, ...detectColorsFromHTML(content)].filter(item => presetColors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
-      convertedContent = convertFromHTML(getFromHTMLConfig({ fontFamilies}))(convertCodeBlock(content))
+      content = content || ''
+      newState.tempColors = [...this.state.tempColors, ...detectColorsFromHTML(content)].filter(item => this.props.colors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
+      convertedContent = convertFromHTML(getFromHTMLConfig({ fontFamilies }))(convertCodeBlock(content))
     } else if (contentFormat === 'raw') {
       convertedContent = convertFromRaw(content)
     }
@@ -299,15 +296,13 @@ export default class BraftEditor extends EditorController {
     } else {
       this.tmpPasteMode = null
     }
-    let { fontFamilies } = this.props
-    fontFamilies = fontFamilies || defaultOptions.fontFamilies
-    const { tempColors } = this.state
+
+    const { fontFamilies } = this.props
     const blockMap = convertFromHTML(getFromHTMLConfig({ fontFamilies }))(convertCodeBlock(html || text)).blockMap
     const nextContentState = Modifier.replaceWithFragment(this.contentState, this.selectionState, blockMap)
-    const presetColors = this.props.colors
 
     this.setState({
-      tempColors: [...tempColors, ...detectColorsFromHTML(html)].filter(item => presetColors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
+      tempColors: [...this.state.tempColors, ...detectColorsFromHTML(html)].filter(item => this.props.colors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
     }, () => {
       this.onChange(EditorState.push(this.editorState, nextContentState, 'insert-fragment'))
     })
@@ -318,10 +313,8 @@ export default class BraftEditor extends EditorController {
 
   addTempColors = (colors, callback) => {
 
-    const presetColors = this.props.colors
-
     this.setState({
-      tempColors: [...this.state.tempColors, ...colors].filter(item => presetColors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
+      tempColors: [...this.state.tempColors, ...colors].filter(item => this.props.colors.indexOf(item) === -1).filter((item, index, array) => array.indexOf(item) === index)
     }, callback)
 
   }
@@ -337,23 +330,11 @@ export default class BraftEditor extends EditorController {
 
     let {
       controls, extendControls, disabled, height, media, language, colors,
-      fontSizes, fontFamilies, emojis, viewWrapper, placeholder, imageControls, lineHeights, letterSpacings, indents, textAlignMaps, needTextBgcolor
+      fontSizes, fontFamilies, emojis, viewWrapper, placeholder, imageControls, lineHeights, letterSpacings, indents, textAlignOptions, allowSetTextBackgroundColor
     } = this.props
 
     const { tempColors } = this.state
-    controls = controls || defaultOptions.controls
-    extendControls = extendControls || defaultOptions.extendControls
     language = languages[language] || languages[defaultOptions.language]
-    colors = colors || defaultOptions.colors
-    fontSizes = fontSizes || defaultOptions.fontSizes
-    fontFamilies = fontFamilies || defaultOptions.fontFamilies
-    lineHeights = lineHeights || defaultOptions.lineHeights
-    emojis = emojis || defaultOptions.emojis
-    height = height || defaultOptions.height
-    letterSpacings = letterSpacings || defaultOptions.letterSpacings
-    textAlignMaps = textAlignMaps || defaultOptions.textAlignMaps
-    needTextBgcolor = needTextBgcolor || defaultOptions.needTextBgcolor
-    indents = indents || defaultOptions.indents
     
     const externalMedias = media && media.externalMedias ? {
       ...defaultOptions.media.externalMedias,
@@ -383,8 +364,8 @@ export default class BraftEditor extends EditorController {
     const controlBarProps = {
       editor: this,
       editorHeight: height,
-      media, controls, language, viewWrapper, extendControls,
-      colors, tempColors, fontSizes, fontFamilies, emojis, lineHeights, letterSpacings, indents, textAlignMaps, needTextBgcolor
+      media, controls, language, viewWrapper, extendControls, colors, tempColors, fontSizes, fontFamilies,
+      emojis, lineHeights, letterSpacings, indents, textAlignOptions, allowSetTextBackgroundColor
     }
 
     const blockRendererFn = getBlockRendererFn({