+++ src/languages/zh-hant.js
@@ -42,6 +42,7 @@ export default {
     code: '代碼',
     link: '鏈接',
     unlink: '清除鏈接',
+    hr: '水平线',
     media: '媒體',
     emoji: '小表情'
   },