+++ src/renderers/styles/blockStyles.js
@@ -12,6 +12,7 @@ export default function blockStyleFn (block) {
   if (blockFloat) {
     result += ` braft-float-${blockFloat}`
   }
+
   return result
 
 }